# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HireAI - AI Interview Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job postings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, closed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/applicants:
    get:
      operationId: getJobApplicants
      tags: [jobs]
      summary: Get ranked applicants for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranked applicants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicantWithScore"
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List interview questions
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create an interview question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, remote]
        status:
          type: string
          enum: [draft, active, closed]
        description:
          type: string
        requirements:
          type: string
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        applicantCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, department, location, type, status, description, requirements, applicantCount, createdAt, updatedAt]
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full_time, part_time, contract, remote]
        status:
          type: string
          enum: [draft, active, closed]
        description:
          type: string
        requirements:
          type: string
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
      required: [title, department, location, type, status, description, requirements]
    Applicant:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        resumeText:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, screening, interviewing, offer, rejected, hired]
        aiScore:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, jobId, name, email, status, appliedAt, updatedAt]
    ApplicantWithScore:
      allOf:
        - $ref: "#/components/schemas/Applicant"
        - type: object
          properties:
            interviewScore:
              type: number
              nullable: true
            interviewsCompleted:
              type: integer
...
    Question:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
          nullable: true
        category:
          type: string
          enum: [ai_collaboration, technical, behavioral, problem_solving, system_design]
        title:
          type: string
        prompt:
          type: string
        aiContext:
          type: string
          description: "Hints on what AI tools/techniques the interviewer expects the candidate to use"
        evaluationCriteria:
          type: string
          description: "What the AI scorer looks for in a great answer"
        difficulty:
          type: string
          enum: [easy, medium, hard]
        timeLimit:
          type: integer
          description: "Time limit in minutes"
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, category, title, prompt, aiContext, evaluationCriteria, difficulty, timeLimit, order, createdAt]
    CreateQuestionBody:
      type: object
      properties:
        jobId:
          type: integer
          nullable: true
        category:
          type: string
          enum: [ai_collaboration, technical, behavioral, problem_solving, system_design]
        title:
          type: string
        prompt:
          type: string
        aiContext:
          type: string
        evaluationCriteria:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        timeLimit:
          type: integer
        order:
          type: integer
      required: [category, title, prompt, aiContext, evaluationCriteria, difficulty, timeLimit, order]
...