# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireAI - AI Interview Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyCategory } from "./createQuestionBodyCategory";
import type { CreateQuestionBodyDifficulty } from "./createQuestionBodyDifficulty";

export interface CreateQuestionBody {
  jobId?: number | null;
  category: CreateQuestionBodyCategory;
  title: string;
  prompt: string;
  aiContext: string;
  evaluationCriteria: string;
  difficulty: CreateQuestionBodyDifficulty;
  timeLimit: number;
  order: number;
}